requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
python-telegram-bot==20.7
tronpy==0.3.7
gunicorn==20.1.0
